+++ src/plugins/intel_cpu/src/nodes/interpolate.cpp
@@ -3,6 +3,7 @@
 //
 
 #include "interpolate.h"
+#include "executors/common/ref_interpolate.hpp"
 
 #include "fake_quantize.h"
 #include "eltwise.h"
@@ -1366,7 +1367,7 @@ private:
 
 namespace {
 struct InterpolateKey {
-    Interpolate::InterpolateAttrs nodeAttrs;
+    InterpolateAttrs nodeAttrs;
     VectorDims srcDims;
     VectorDims dstDims;
     std::vector<float> dataScales;
@@ -1452,29 +1453,6 @@ inline SizeVector getBlockND(const SizeVector& shape) {
         blockND[i] = shape[i] * blockND[i+1];
     }
     return blockND;
-}
-// w/hw/ncw/nchw/ncdhw to ncdhw
-inline SizeVector to5Dim(SizeVector casesDim) {
-    size_t caseSize = casesDim.size();
-    SizeVector dim5(5, 1lu);
-    dim5[4] = casesDim[caseSize - 1];
-    if (caseSize > 1) {
-        dim5[3] = casesDim[caseSize - 2];
-    }
-    if (caseSize > 2) {
-        dim5[0] = casesDim[0];
-    }
-    if (caseSize > 3) {
-        dim5[1] = casesDim[1];
-    }
-    if (caseSize > 4) {
-        dim5[2] = casesDim[2];
-    }
-    if (caseSize == 3) {  // nhw -> ncw
-        dim5[1] = dim5[3];
-        dim5[3] = 1lu;
-    }
-    return dim5;
 }
 
 using ngInterpMode = ngraph::opset4::Interpolate::InterpolateMode;
@@ -1954,10 +1932,6 @@ void Interpolate::createPrimitive() {
     }
 }
 
-inline int clipCoord(int pos, int length) {
-    return std::max(static_cast<int>(0), std::min(pos, length - 1));
-}
-
 static inline float triangleCoeff(float x) {
     return (std::max)(0.0f, 1 - std::abs(x));
 }
@@ -1983,17 +1957,6 @@ void Interpolate::setPostOps(dnnl::primitive_attr &attr, const VectorDims &dims)
     }
 
     attr.set_post_ops(ops);
-}
-
-SizeVector Interpolate::getPaddedInputShape(const VectorDims &srcDims,
-                                                      const std::vector<int> &padBegin,
-                                                      const std::vector<int> &padEnd) {
-    SizeVector paddedShape;
-    int dataRank = srcDims.size();
-    for (int i = 0; i < dataRank; i++) {
-        paddedShape.push_back(srcDims[i] + padBegin[i] + padEnd[i]);
-    }
-    return paddedShape;
 }
 
 // get scales of data rank size
@@ -2366,763 +2329,6 @@ void Interpolate::InterpolateJitExecutor::cubicPlanar(const uint8_t *in_ptr_, ui
     });
 }
 
-// =====================================================================================================================
-// index layout:
-// d_0............d_OD-1, h_0..............h_OH-1, w_0................w_OW-1
-void Interpolate::InterpolateExecutor::buildTblNN(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d,
-                                        const std::vector<float>& dataScales, InterpolateLayoutType layout, InterpolateNearestMode nearestMode) {
-    const int dimSize = dataRank;
-    float fz = (dimSize == 5) ? dataScales[dimSize - 3] : 1.f;
-    float fy = dataScales[dimSize - 2];
-    float fx = dataScales[dimSize - 1];
-    size_t ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
-
-    indexTable.resize(OD + OH + OW);
-    bool isDDownsample = (fz < 1) ? true : false;
-    bool isHDownsample = (fy < 1) ? true : false;
-    bool isWDownsample = (fx < 1) ? true : false;
-    for (int oz = 0; oz < OD; oz++) {
-        float iz = coordTransToInput(oz, fz, ID, OD);
-        indexTable[oz] = nearestRound(iz, isDDownsample, nearestMode);
-        indexTable[oz] = clipCoord(indexTable[oz], ID);
-    }
-    for (int oy = 0; oy < OH; oy++) {
-        float iy = coordTransToInput(oy, fy, IH, OH);
-        indexTable[OD + oy] = nearestRound(iy, isHDownsample, nearestMode);
-        indexTable[OD + oy] = clipCoord(indexTable[OD + oy], IH);
-    }
-    for (int ox = 0; ox < OW; ox++) {
-        float ix = coordTransToInput(ox, fx, IW, OW);
-        indexTable[OD + OH + ox] = nearestRound(ix, isWDownsample, nearestMode);
-        indexTable[OD + OH + ox] = clipCoord(indexTable[OD + OH + ox], IW);
-    }
-}
-
-// scale is float(outShape) / float(inShape)
-// strictly consistent with onnx calc manner(div scale, not multiply inverse), given this is done offline
-// the slight precison diff can produce obvious wrong value due to "nearest round" behavior for NN mode
-float Interpolate::InterpolateExecutor::coordTransToInput(int outCoord, float scale, int inShape, int outShape) const {
-    if (scale == 1.0f || (inShape == outShape)) {
-        return outCoord;
-    }
-    switch (coordTransMode) {
-        case InterpolateCoordTransMode::half_pixel: {
-            return (outCoord + 0.5f) / scale - 0.5f;
-            break;
-        }
-        case InterpolateCoordTransMode::pytorch_half_pixel: {
-            if (outShape > 1)
-                return (outCoord + 0.5f) / scale - 0.5f;
-            else
-                return 0;
-            break;
-        }
-        case InterpolateCoordTransMode::asymmetric: {
-            return static_cast<float>(outCoord) / scale;
-            break;
-        }
-        case InterpolateCoordTransMode::tf_half_pixel_for_nn: {
-            return (outCoord + 0.5f) / scale;
-            break;
-        }
-        case InterpolateCoordTransMode::align_corners: {
-            if (outShape > 1)
-                return outCoord * (static_cast<float>(inShape - 1) / static_cast<float>(outShape - 1));
-            else
-                return 0;
-            break;
-        }
-        default: {
-            IE_THROW() << "errorPrefix" << " does not support specified coordinate transformation mode";
-            break;
-        }
-    }
-}
-
-int Interpolate::InterpolateExecutor::nearestRound(float originCoord, bool isDownsample, InterpolateNearestMode nearestMode) const {
-    switch (nearestMode) {
-        case InterpolateNearestMode::round_prefer_floor: {
-            if (originCoord == (static_cast<int>(originCoord) + 0.5f))
-                return static_cast<int>(std::floor(originCoord));
-            else
-                return static_cast<int>(std::round(originCoord));
-            break;
-        }
-        case InterpolateNearestMode::round_prefer_ceil: {
-            return static_cast<int>(std::round(originCoord));
-            break;
-        }
-        case InterpolateNearestMode::floor: {
-            return static_cast<int>(std::floor(originCoord));
-            break;
-        }
-        case InterpolateNearestMode::ceil: {
-            return static_cast<int>(std::ceil(originCoord));
-            break;
-        }
-        case InterpolateNearestMode::simple: {
-            if (isDownsample)
-                return static_cast<int>(std::ceil(originCoord));
-            else
-                return static_cast<int>(originCoord);
-        }
-        default: {
-            IE_THROW() << "errorPrefix" << " does not support specified nearest round mode";
-            break;
-        }
-    }
-}
-
-void Interpolate::InterpolateExecutor::linearOnnxCF(int outCoord, float scale, int inShape, int outShape,
-                int& index0, int& index1, float& weight0, float& weight1) {
-    float inCoord = coordTransToInput(outCoord, scale, inShape, outShape);
-    inCoord = std::max(0.0f, std::min(inCoord, static_cast<float>(inShape - 1)));
-    index0 = std::min(static_cast<int>(inCoord), inShape - 1);
-    index1 = std::min(index0 + 1, inShape - 1);
-
-    weight1 = std::fabs(inCoord - index0);
-    weight0 = std::fabs(inCoord - index1);
-    if (index0 == index1) {
-        weight0 = 0.5f;
-        weight1 = 0.5f;
-    }
-}
-
-void Interpolate::InterpolateExecutor::buildTblLinearOnnx(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d,
-                                                const std::vector<float>& dataScales, InterpolateLayoutType layout) {
-    int dimSize = dataRank;
-    float fz = (spatialDimSize > 2) ? dataScales[dimSize - 3] : 1.f;
-    float fy = (spatialDimSize > 1) ? dataScales[dimSize - 2] : 1.f;
-    float fx = dataScales[dimSize - 1];
-    int ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    int OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
-
-    std::vector<int*> indexPtr(MAX_INPUT_INTERPOLATE, 0);
-    std::vector<float*> weightPtr(MAX_INPUT_INTERPOLATE, 0);
-    if (layout == InterpolateLayoutType::planar) {
-        // FrontTopLeft:0, FrontTopRight:1, FrontBottomLeft:2, FrontBottomRight:3,
-        // EndTopLeft:4,   EndTopRight:5,   EndBottomLeft:6,   EndBottomRight:7
-        // weight: Left:0, ritht:1, top:2, bottom:3, front:4, end:5
-        int eltInGrid = (spatialDimSize > 2) ? MAX_INPUT_INTERPOLATE : ((spatialDimSize > 1) ? 4 : 2);
-        int idxType = 2;
-        int scratchLen = rnd_up(eltInGrid * OW * OH * OD, 16);
-        indexTable.resize(idxType * scratchLen);
-
-        indexPtr[0] = static_cast<int*>(&indexTable[0]);
-        indexPtr[1] = static_cast<int*>(&indexTable[OW * OH * OD]);
-        weightPtr[0] = reinterpret_cast<float*>(&indexTable[scratchLen]);
-        weightPtr[1] = reinterpret_cast<float*>(&indexTable[scratchLen + OW * OH * OD]);
-        if (spatialDimSize > 1) {
-            indexPtr[2] = static_cast<int*>(&indexTable[2 * OW * OH * OD]);
-            indexPtr[3] = static_cast<int*>(&indexTable[3 * OW * OH * OD]);
-            weightPtr[2] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW * OH * OD]);
-            weightPtr[3] = reinterpret_cast<float*>(&indexTable[scratchLen + 3 * OW * OH * OD]);
-        }
-        if (spatialDimSize > 2) {
-            indexPtr[4] = static_cast<int*>(&indexTable[4 * OW * OH * OD]);
-            indexPtr[5] = static_cast<int*>(&indexTable[5 * OW * OH * OD]);
-            indexPtr[6] = static_cast<int*>(&indexTable[6 * OW * OH * OD]);
-            indexPtr[7] = static_cast<int*>(&indexTable[7 * OW * OH * OD]);
-            weightPtr[4] = reinterpret_cast<float*>(&indexTable[scratchLen + 4 * OW * OH * OD]);
-            weightPtr[5] = reinterpret_cast<float*>(&indexTable[scratchLen + 5 * OW * OH * OD]);
-        }
-        int scale = mayiuse(cpu::x64::sse41) ? srcDataSize : 1;
-
-        for (int oz = 0; oz < OD; oz++) {
-            int izF, izE;
-            float weightF, weightE;
-            linearOnnxCF(oz, fz, ID, OD, izF, izE, weightF, weightE);
-            int idxOz = oz * OH * OW;
-            for (int oy = 0; oy < OH; oy++) {
-                int iyT, iyB;
-                float weightT, weightB;
-                linearOnnxCF(oy, fy, IH, OH, iyT, iyB, weightT, weightB);
-                int idxOzOy = idxOz + oy * OW;
-                for (int ox = 0; ox < OW; ox++) {
-                    int ixL, ixR;
-                    float weightL, weightR;
-                    linearOnnxCF(ox, fx, IW, OW, ixL, ixR, weightL, weightR);
-
-                    int idxOzOyOx = idxOzOy + ox;
-                    indexPtr[0][idxOzOyOx] = (izF * IH * IW + iyT * IW + ixL) * scale;
-                    indexPtr[1][idxOzOyOx] = (izF * IH * IW + iyT * IW + ixR) * scale;
-                    weightPtr[0][idxOzOyOx] = weightL;
-                    weightPtr[1][idxOzOyOx] = weightR;
-                    if (spatialDimSize  > 1) {
-                        indexPtr[2][idxOzOyOx] = (izF * IH * IW + iyB * IW + ixL) * scale;
-                        indexPtr[3][idxOzOyOx] = (izF * IH * IW + iyB * IW + ixR) * scale;
-                        weightPtr[2][idxOzOyOx] = weightT;
-                        weightPtr[3][idxOzOyOx] = weightB;
-                    }
-                    if (spatialDimSize > 2) {
-                        indexPtr[4][idxOzOyOx] = (izE * IH * IW + iyT * IW + ixL) * scale;
-                        indexPtr[5][idxOzOyOx] = (izE * IH * IW + iyT * IW + ixR) * scale;
-                        indexPtr[6][idxOzOyOx] = (izE * IH * IW + iyB * IW + ixL) * scale;
-                        indexPtr[7][idxOzOyOx] = (izE * IH * IW + iyB * IW + ixR) * scale;
-                        weightPtr[4][idxOzOyOx] = weightF;
-                        weightPtr[5][idxOzOyOx] = weightE;
-                    }
-                }
-            }
-        }
-    } else {
-        // index: left:OW right:OW Top:OH Bottom:OH, Front:OD, End:OD
-        // weight:same as index
-        size_t scratchLen = rnd_up(OW + OW + OH + OH + OD + OD, 16);
-        int idxType = 2;
-        indexTable.resize(idxType * scratchLen);
-        indexPtr[0] = static_cast<int*>(&indexTable[0]);
-        indexPtr[1] = static_cast<int*>(&indexTable[OW]);
-        indexPtr[2] = static_cast<int*>(&indexTable[2 * OW]);
-        indexPtr[3] = static_cast<int*>(&indexTable[2 * OW + OH]);
-        indexPtr[4] = static_cast<int*>(&indexTable[2 * OW + 2 * OH]);
-        indexPtr[5] = static_cast<int*>(&indexTable[2 * OW + 2 * OH + OD]);
-
-        weightPtr[0] = reinterpret_cast<float*>(&indexTable[scratchLen]);
-        weightPtr[1] = reinterpret_cast<float*>(&indexTable[scratchLen + OW]);
-        weightPtr[2] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW]);
-        weightPtr[3] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + OH]);
-        weightPtr[4] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + 2 * OH]);
-        weightPtr[5] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW + 2 * OH + OD]);
-
-        for (int ox = 0; ox < OW; ox++) {
-            linearOnnxCF(ox, fx, IW, OW, indexPtr[0][ox], indexPtr[1][ox], weightPtr[0][ox], weightPtr[1][ox]);
-        }
-        for (int oy = 0; oy < OH; oy++) {
-            linearOnnxCF(oy, fy, IH, OH, indexPtr[2][oy], indexPtr[3][oy], weightPtr[2][oy], weightPtr[3][oy]);
-        }
-        for (int oz = 0; oz < OD; oz++) {
-            linearOnnxCF(oz, fz, ID, OD, indexPtr[4][oz], indexPtr[5][oz], weightPtr[4][oz], weightPtr[5][oz]);
-        }
-    }
-}
-
-// table layout:
-// wd .........wd, wh............wh, ww.............ww, id...........id, ih............ih, iw..............iw
-//                        |                                                      |
-//                   wh0.....wh_diameter                                    ih0.....ih_diameter
-void Interpolate::InterpolateExecutor::buildTblLinear(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d,
-                                            const std::vector<float>& dataScales, int kernel_width, bool antialias) {
-    int dimSize = dataRank;
-    float fz = (dimSize == 5) ? dataScales[dimSize - 3] : 1.f;
-    float fy = dataScales[dimSize - 2];
-    float fx = dataScales[dimSize - 1];
-    size_t ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
-
-    if (!(IW == OW && IH == OH && ID == OD)) {
-        float ax = antialias ? fx : 1.0f;
-        float ay = antialias ? fy : 1.0f;
-        float az = antialias ? fz : 1.0f;
-
-        int rx = (fx > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / ax));
-        int ry = (fy > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / ay));
-        int rz = (fz > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / az));
-
-        int diaOD = 2 * rz + 1;
-        int diaOH = 2 * ry + 1;
-        int diaOW = 2 * rx + 1;
-        int sizeOD = OD * diaOD;
-        int sizeOH = OH * diaOH;
-        int sizeOW = OW * diaOW;
-        indexTable.resize((sizeOD + sizeOH + sizeOW) * 2);
-        float *weightTable = reinterpret_cast<float*>(&indexTable[0]);
-        float *weightOD = static_cast<float*>(&weightTable[0]);
-        float *weightOH = static_cast<float*>(&weightTable[sizeOD]);
-        float *weightOW = static_cast<float*>(&weightTable[sizeOD + sizeOH]);
-
-        int *idxTable = static_cast<int*>(&indexTable[sizeOD + sizeOH + sizeOW]);
-        int *idxOD = static_cast<int*>(&idxTable[0]);
-        int *idxOH = static_cast<int*>(&idxTable[sizeOD]);
-        int *idxOW = static_cast<int*>(&idxTable[sizeOD + sizeOH]);
-
-        for (int oz = 0; oz < OD; oz++) {
-            float iz = coordTransToInput(oz, fz, ID, OD);
-            int iz_r = static_cast<int>(std::round(iz));
-            for (int r = iz_r - rz, i = 0; r <= iz_r + rz; r++, i++) {
-                idxOD[oz * diaOD + i] = r;
-                if (r < 0 || r >= static_cast<int>(ID)) {
-                    weightOD[oz * diaOD + i] = 0.f;
-                } else {
-                    float dz = iz - r;
-                    weightOD[oz * diaOD + i] = az * triangleCoeff(az * dz);
-                }
-            }
-        }
-        for (int oy = 0; oy < OH; oy++) {
-            float iy = coordTransToInput(oy, fy, IH, OH);
-            int iy_r = static_cast<int>(std::round(iy));
-            for (int r = iy_r - ry, i = 0; r <= iy_r + ry; r++, i++) {
-                idxOH[oy * diaOH + i] = r;
-                if (r < 0 || r >= static_cast<int>(IH)) {
-                    weightOH[oy * diaOH + i] = 0.f;
-                } else {
-                    float dy = iy - r;
-                    weightOH[oy * diaOH + i] = ay * triangleCoeff(ay * dy);
-                }
-            }
-        }
-        for (int ox = 0; ox < OW; ox++) {
-            float ix = coordTransToInput(ox, fx, IW, OW);
-            int ix_r = static_cast<int>(std::round(ix));
-            for (int r = ix_r - rx, i = 0; r <= ix_r + rx; r++, i++) {
-                idxOW[ox * diaOW + i] = r;
-                if (r < 0 || r >= static_cast<int>(IW)) {
-                    weightOW[ox * diaOW + i] = 0.f;
-                } else {
-                    float dx = ix - r;
-                    weightOW[ox * diaOW + i] = ax * triangleCoeff(ax * dx);
-                }
-            }
-        }
-    }
-}
-
-std::vector<float> Interpolate::InterpolateExecutor::getCubicCoeffs(float mantissa, float a) {
-    float m = std::fabs(mantissa);
-    std::vector<float> coeffs(4, 0.f);
-
-    coeffs[0] = a * (m - 1.0) * (m - 1.0) * m;
-    coeffs[1] = ((a + 2.0) * m - (a + 3.0)) * m * m + 1.0;
-    coeffs[2] = (((-a - 2.0) * m + (2.0 * a + 3.0)) * m - a) * m;
-    coeffs[3] = -a * m * m * (m - 1.0);
-    return coeffs;
-}
-
-// table layout:
-// OW      OW         OW         OW         OW          OH       OH           OH           OH           OH
-// x_idx   x_weight0  x_weight1  x_weight2  x_weight3   y_idx    y_weight0    y_weight1    y_weight2    y_weight3
-void Interpolate::InterpolateExecutor::buildTblCubic(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d, const std::vector<float>& dataScales,
-                                        float cubicCoeff, InterpolateLayoutType layout) {
-    int dimSize = dataRank;
-    float fy = dataScales[dimSize - 2];
-    float fx = dataScales[dimSize - 1];
-    int IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    int OH = dstDim5d[3], OW = dstDim5d[4];
-
-    // idxNum for index, CUBIC_GRID_LEN for weight
-    const int idxNum = 1;
-    size_t idxWeightSize = (CUBIC_GRID_LEN + idxNum) * OW + (CUBIC_GRID_LEN + idxNum) * OH;
-    if (layout != InterpolateLayoutType::planar) {
-        indexTable.resize(idxWeightSize);
-    } else {
-        size_t sequenceSize = 2 * OH * OW;
-        indexTable.resize(idxWeightSize + sequenceSize);
-    }
-
-    int tblAdvance = 0;
-    int *xOrigin = static_cast<int*>(&indexTable[tblAdvance]);
-    tblAdvance += OW;
-    float *xFactor = reinterpret_cast<float*>(&indexTable[tblAdvance]);
-    for (int ox = 0; ox < OW; ox++) {
-        float ix = coordTransToInput(ox, fx, IW, OW);
-        int ix_r = static_cast<int>(std::floor(ix));
-        xOrigin[ox] = ix_r;
-        float m = ix - ix_r;
-        std::vector<float> coffes = getCubicCoeffs(m, cubicCoeff);
-        xFactor[CUBIC_GRID_LEN * ox] = coffes[0];
-        xFactor[CUBIC_GRID_LEN * ox + 1] = coffes[1];
-        xFactor[CUBIC_GRID_LEN * ox + 2] = coffes[2];
-        xFactor[CUBIC_GRID_LEN * ox + 3] = coffes[3];
-    }
-
-    tblAdvance += CUBIC_GRID_LEN * OW;
-    int *yOrigin = static_cast<int*>(&indexTable[tblAdvance]);
-    tblAdvance += OH;
-    float *yFactor = reinterpret_cast<float*>(&indexTable[tblAdvance]);
-    for (int oy = 0; oy < OH; oy++) {
-        float iy = coordTransToInput(oy, fy, IH, OH);
-        int iy_r = static_cast<int>(std::floor(iy));
-        yOrigin[oy] = iy_r;
-        float m = iy - iy_r;
-        std::vector<float> coffes = getCubicCoeffs(m, cubicCoeff);
-        yFactor[CUBIC_GRID_LEN * oy] = coffes[0];
-        yFactor[CUBIC_GRID_LEN * oy + 1] = coffes[1];
-        yFactor[CUBIC_GRID_LEN * oy + 2] = coffes[2];
-        yFactor[CUBIC_GRID_LEN * oy + 3] = coffes[3];
-    }
-
-    if (layout == InterpolateLayoutType::planar) {
-        tblAdvance += CUBIC_GRID_LEN * OH;
-        int *sequenceOH = static_cast<int*>(&indexTable[tblAdvance]);
-        tblAdvance += OH * OW;
-        int *sequenceOW = static_cast<int*>(&indexTable[tblAdvance]);
-        for (int h = 0; h < OH; ++h) {
-            int offset = h * OW;
-            for (int w = 0; w < OW; ++w) {
-                sequenceOH[offset + w] = h * sizeof(int);
-                sequenceOW[offset + w] = w * sizeof(int);
-            }
-        }
-    }
-}
-
-void Interpolate::InterpolateRefExecutor::NNRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW,
-                                                          int OD, int OH, int OW) {
-    int *index_d = static_cast<int*>(&indexTable[0]);
-    int *index_h = static_cast<int*>(&indexTable[OD]);
-    int *index_w = static_cast<int*>(&indexTable[OD + OH]);
-
-    const float *in_ptr_f32 = reinterpret_cast<const float *>(in_ptr_);
-    float *out_ptr_f32 = reinterpret_cast<float *>(out_ptr_);
-
-    parallel_for3d(B, C, OD, [&](size_t b, size_t c, size_t od) {
-        const float *in_ptr = in_ptr_f32 + (IW * IH * ID * C * b + IW * IH * ID * c + IW * IH * index_d[od]);
-        float *out_ptr = out_ptr_f32 + (OW * OH * OD * C * b + OW * OH * OD * c + OW * OH * od);
-        for (int oh = 0; oh < OH; oh++) {
-            const float *in_ptr_h = in_ptr + (IW * index_h[oh]);
-            float *out_ptr_h = out_ptr + (OW * oh);
-            for (int ow = 0; ow < OW; ow++) {
-                out_ptr_h[ow] = in_ptr_h[index_w[ow]];
-            }
-        }
-    });
-}
-
-void Interpolate::InterpolateRefExecutor::linearOnnxRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW,
-                                                                  int OD, int OH, int OW) {
-    std::vector<int*> indexPtr(MAX_INPUT_INTERPOLATE, 0);
-    std::vector<float*> weightPtr(MAX_INPUT_INTERPOLATE, 0);
-    // FrontTopLeft:0, FrontTopRight:1, FrontBottomLeft:2, FrontBottomRight:3,
-    // EndTopLeft:4,   EndTopRight:5,   EndBottomLeft:6,   EndBottomRight:7
-    // weight: Left:0, ritht:1, top:2, bottom:3, front:4, end:5
-
-    int eltInGrid = (spatialDimSize > 2) ? MAX_INPUT_INTERPOLATE : ((spatialDimSize > 1) ? 4 : 2);
-    int scratchLen = rnd_up(eltInGrid * OW * OH * OD, 16);
-
-    indexPtr[0] = static_cast<int*>(&indexTable[0]);
-    indexPtr[1] = static_cast<int*>(&indexTable[OW * OH * OD]);
-    weightPtr[0] = reinterpret_cast<float*>(&indexTable[scratchLen]);
-    weightPtr[1] = reinterpret_cast<float*>(&indexTable[scratchLen + OW * OH * OD]);
-    if (spatialDimSize > 1) {
-        indexPtr[2] = static_cast<int*>(&indexTable[2 * OW * OH * OD]);
-        indexPtr[3] = static_cast<int*>(&indexTable[3 * OW * OH * OD]);
-        weightPtr[2] = reinterpret_cast<float*>(&indexTable[scratchLen + 2 * OW * OH * OD]);
-        weightPtr[3] = reinterpret_cast<float*>(&indexTable[scratchLen + 3 * OW * OH * OD]);
-    }
-    if (spatialDimSize > 2) {
-        indexPtr[4] = static_cast<int*>(&indexTable[4 * OW * OH * OD]);
-        indexPtr[5] = static_cast<int*>(&indexTable[5 * OW * OH * OD]);
-        indexPtr[6] = static_cast<int*>(&indexTable[6 * OW * OH * OD]);
-        indexPtr[7] = static_cast<int*>(&indexTable[7 * OW * OH * OD]);
-        weightPtr[4] = reinterpret_cast<float*>(&indexTable[scratchLen + 4 * OW * OH * OD]);
-        weightPtr[5] = reinterpret_cast<float*>(&indexTable[scratchLen + 5 * OW * OH * OD]);
-    }
-
-    const float *in_ptr_f32 = reinterpret_cast<const float *>(in_ptr_);
-    float *out_ptr_f32 = reinterpret_cast<float *>(out_ptr_);
-
-    parallel_for2d(B, C, [&](size_t b, size_t c) {
-        float *out_ptr_nc = out_ptr_f32 + (OD * OH * OW * C * b + OD * OH * OW * c);
-        const float *in_ptr_nc = in_ptr_f32 + (ID * IH * IW * C * b + ID * IH * IW * c);
-        // do not combined 1d/2d to 3d unified process to get rid of invalid computing.
-        switch (spatialDimSize) {
-            case 1:
-                for (int i = 0; i < OW; i++) {
-                    float src0 = in_ptr_nc[indexPtr[0][i]];
-                    float src1 = in_ptr_nc[indexPtr[1][i]];
-
-                    out_ptr_nc[i] = src0 * weightPtr[0][i] +
-                                    src1 * weightPtr[1][i];
-                }
-                break;
-            case 2:
-                for (int i = 0; i < OH * OW; i++) {
-                    float src00 = in_ptr_nc[indexPtr[0][i]];
-                    float src01 = in_ptr_nc[indexPtr[1][i]];
-                    float src10 = in_ptr_nc[indexPtr[2][i]];
-                    float src11 = in_ptr_nc[indexPtr[3][i]];
-
-                    out_ptr_nc[i] = src00 * weightPtr[2][i] * weightPtr[0][i] +
-                                    src01 * weightPtr[2][i] * weightPtr[1][i] +
-                                    src10 * weightPtr[3][i] * weightPtr[0][i] +
-                                    src11 * weightPtr[3][i] * weightPtr[1][i];
-                }
-                break;
-            case 3:
-                for (int i = 0; i < OD * OH * OW; i++) {
-                    float src000 = in_ptr_nc[indexPtr[0][i]];
-                    float src001 = in_ptr_nc[indexPtr[1][i]];
-                    float src010 = in_ptr_nc[indexPtr[2][i]];
-                    float src011 = in_ptr_nc[indexPtr[3][i]];
-                    float src100 = in_ptr_nc[indexPtr[4][i]];
-                    float src101 = in_ptr_nc[indexPtr[5][i]];
-                    float src110 = in_ptr_nc[indexPtr[6][i]];
-                    float src111 = in_ptr_nc[indexPtr[7][i]];
-
-                    // float dstValue =
-                    // weightPtr[4][i] * weightPtr[2][i] * weightPtr[0][i] * src000 +
-                    // weightPtr[4][i] * weightPtr[2][i] * weightPtr[1][i] * src001 +
-                    // weightPtr[4][i] * weightPtr[3][i] * weightPtr[0][i] * src010 +
-                    // weightPtr[4][i] * weightPtr[3][i] * weightPtr[1][i] * src011 +
-                    // weightPtr[5][i] * weightPtr[2][i] * weightPtr[0][i] * src100 +
-                    // weightPtr[5][i] * weightPtr[2][i] * weightPtr[1][i] * src101 +
-                    // weightPtr[5][i] * weightPtr[3][i] * weightPtr[0][i] * src110 +
-                    // weightPtr[5][i] * weightPtr[3][i] * weightPtr[1][i] * src111;
-
-                    out_ptr_nc[i] =
-                    weightPtr[4][i] * (weightPtr[2][i] * (weightPtr[0][i] * src000 +
-                                                          weightPtr[1][i] * src001) +
-                                       weightPtr[3][i] * (weightPtr[0][i] * src010 +
-                                                          weightPtr[1][i] * src011)) +
-                    weightPtr[5][i] * (weightPtr[2][i] * (weightPtr[0][i] * src100 +
-                                                          weightPtr[1][i] * src101) +
-                                       weightPtr[3][i] * (weightPtr[0][i] * src110 +
-                                                          weightPtr[1][i] * src111));
-                }
-                break;
-            default:
-                break;
-        }
-    });
-}
-
-void Interpolate::InterpolateRefExecutor::cubicRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int IH, int IW, int OH, int OW) {
-    const int idxNum = 1;
-    int *xOrigin = static_cast<int*>(&indexTable[0]);
-    float *xFactor = reinterpret_cast<float*>(&indexTable[OW]);
-    int *yOrigin = static_cast<int*>(&indexTable[(CUBIC_GRID_LEN + idxNum) * OW]);
-    float *yFactor = reinterpret_cast<float*>(&indexTable[(CUBIC_GRID_LEN + idxNum) * OW + OH]);
-
-    const float *in_ptr_f32 = reinterpret_cast<const float *>(in_ptr_);
-    float *out_ptr_f32 = reinterpret_cast<float *>(out_ptr_);
-
-    parallel_for4d(B, C, OH, OW, [&](size_t n, size_t c, size_t oy, size_t ox) {
-        const float *in_ptr_nc = in_ptr_f32 + (IW * IH * C * n + IW * IH * c);
-        float *out_ptr_nc = out_ptr_f32 + (OW * OH * C * n + OW * OH * c);
-
-        int iy = yOrigin[oy];
-        int ix = xOrigin[ox];
-
-        float retY = 0.f;
-        for (int y = iy - 1, i = 0; y <= iy + 2; y++, i++) {
-            int yInRange = std::max(0, std::min(y, IH - 1));
-            const float *in_ptr_nch = in_ptr_nc + IW * yInRange;
-            float retX = 0.f;
-            for (int x = ix - 1, j = 0; x <= ix + 2; x++, j++) {
-                int xInRange = std::max(0, std::min(x, IW - 1));
-                retX += xFactor[ox * CUBIC_GRID_LEN + j] * in_ptr_nch[xInRange];
-            }
-            retY += yFactor[oy * CUBIC_GRID_LEN + i] * retX;
-        }
-        out_ptr_nc[oy * OW + ox] = retY;
-    });
-}
-
-float Interpolate::InterpolateRefExecutor::getValue(const uint8_t *base, size_t offset, InferenceEngine::Precision prec) {
-    const uint8_t *baseOffset = base + offset;
-    switch (prec) {
-        case Precision::U8: {
-            return static_cast<float>(*baseOffset);
-            break;
-        }
-        case Precision::I8: {
-            const int8_t *valuePtr = reinterpret_cast<const int8_t *>(baseOffset);
-            return static_cast<float>(*valuePtr);
-            break;
-        }
-        case Precision::BF16: {
-            const uint16_t *valuePtr = reinterpret_cast<const uint16_t *>(baseOffset);
-            return bfloat16_t::from_bits(*valuePtr);
-            break;
-        }
-        case Precision::FP32: {
-            const float *valuePtr = reinterpret_cast<const float *>(baseOffset);
-            return *valuePtr;
-            break;
-        }
-        default: {
-            IE_THROW() << "Interpolate layer does not support precision: " << prec;
-            break;
-        }
-    }
-}
-
-void Interpolate::InterpolateRefExecutor::setValue(uint8_t *base, size_t offset, float value, InferenceEngine::Precision prec) {
-    uint8_t *baseOffset = base + offset;
-    switch (prec) {
-        case Precision::U8: {
-            uint8_t data = static_cast<uint8_t>(value < 0 ? 0 : value);
-            cpu_memcpy(baseOffset, &data, 1);
-            break;
-        }
-        case Precision::I8: {
-            int8_t data = static_cast<int8_t>(value);
-            cpu_memcpy(baseOffset, &data, 1);
-            break;
-        }
-        case Precision::BF16: {
-            uint16_t data = bfloat16_t(value).to_bits();
-            cpu_memcpy(baseOffset, &data, 2);
-            break;
-        }
-        case Precision::FP32: {
-            cpu_memcpy(baseOffset, &value, sizeof(float));
-            break;
-        }
-        default: {
-            IE_THROW() << "Interpolate layer does not support precision: " << prec;
-            break;
-        }
-    }
-}
-
-void Interpolate::InterpolateRefExecutor::linearInterpolation(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW,
-                                          float fx, float fy, float fz, int OD, int OH, int OW, int kernel_width, bool antialias) {
-    if (IW == OW && IH == OH && ID == OD) {
-        size_t spatialDimSize = IW * IH * ID;
-        // TODO: enable when fusing into interp with linear mode will support
-        if (/*fusedWith.empty() &&*/ inputPrec == outputPrec) {
-            size_t size = B * C * spatialDimSize * srcDataSize;
-            cpu_memcpy(out_ptr_, in_ptr_, size);
-        } else {
-            parallel_for2d(B, C, [&](size_t b, size_t c) {
-                const uint8_t *in_ptr_nc = in_ptr_ + (spatialDimSize * C * b + spatialDimSize * c) * srcDataSize;
-                uint8_t *out_ptr_nc = out_ptr_ + (spatialDimSize * C * b + spatialDimSize * c) * dstDataSize;
-                for (size_t i = 0; i < spatialDimSize; i++) {
-                    float dstValue = getValue(in_ptr_nc, i * srcDataSize, inputPrec);
-                    setValue(out_ptr_nc, i * dstDataSize, dstValue, outputPrec);
-                }
-            });
-        }
-        return;
-    }
-
-    float ax = antialias ? fx : 1.0f;
-    float ay = antialias ? fy : 1.0f;
-    float az = antialias ? fz : 1.0f;
-
-    int rx = (fx > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / ax));
-    int ry = (fy > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / ay));
-    int rz = (fz > 1.0f) ? 2 : static_cast<int>(ceil(static_cast<float>(kernel_width) / az));
-
-    int diaOD = 2 * rz + 1;
-    int diaOH = 2 * ry + 1;
-    int diaOW = 2 * rx + 1;
-    int sizeOD = OD * diaOD;
-    int sizeOH = OH * diaOH;
-    int sizeOW = OW * diaOW;
-
-    float *weightTable = reinterpret_cast<float*>(&indexTable[0]);
-    float *weightOD = static_cast<float*>(&weightTable[0]);
-    float *weightOH = static_cast<float*>(&weightTable[sizeOD]);
-    float *weightOW = static_cast<float*>(&weightTable[sizeOD + sizeOH]);
-
-    int *idxTable = static_cast<int*>(&indexTable[sizeOD + sizeOH + sizeOW]);
-    int *idxOD = static_cast<int*>(&idxTable[0]);
-    int *idxOH = static_cast<int*>(&idxTable[sizeOD]);
-    int *idxOW = static_cast<int*>(&idxTable[sizeOD + sizeOH]);
-
-    parallel_for2d(B, C, [&](size_t b, size_t c) {
-        const uint8_t *in_ptr_nc = in_ptr_ + (IW * IH * ID * C * b + IW * IH * ID * c) * srcDataSize;
-        uint8_t *out_ptr_nc = out_ptr_ + (OW * OH * OD * C * b + OW * OH * OD * c) * dstDataSize;
-        for (size_t oz = 0; oz < OD; oz++) {
-            uint8_t *out_ptr_ncd = out_ptr_nc + (OW * OH * oz) * dstDataSize;
-            for (size_t oy = 0; oy < OH; oy++) {
-                uint8_t *out_ptr_ncdh = out_ptr_ncd + (OW * oy) * dstDataSize;
-                for (size_t ox = 0; ox < OW; ox++) {
-                    float sum = 0.f;
-                    float wsum = 0.f;
-
-                    // this comment explains the original algo.
-                    // for (int z = iz_r - rz; z <= iz_r + rz; z++) {
-                    //    for (int y = iy_r - ry; y <= iy_r + ry; y++) {
-                    //        for (int x = ix_r - rx; x <= ix_r + rx; x++) {
-                    //            bool is_continue =  z < 0                     ||
-                    //                                y < 0                     ||
-                    //                                x < 0                     ||
-                    //                                z >= static_cast<int>(ID) ||
-                    //                                y >= static_cast<int>(IH) ||
-                    //                                x >= static_cast<int>(IW);
-                    //            if (is_continue)
-                    //                continue;
-
-                    //            float dx = ix - x;
-                    //            float dy = iy - y;
-                    //            float dz = iz - z;
-
-                    //            float w = ax * triangleCoeff(ax * dx) *
-                    //                      ay * triangleCoeff(ay * dy) *
-                    //                      az * triangleCoeff(az * dz);
-
-                    //            sum += w * getValue(in_ptr_nc, (z * IH * IW + y * IW + x) * srcDataSize, inputPrec);
-                    //            wsum += w;
-                    //        }
-                    //    }
-                    //}
-
-                    for (int iz = 0; iz < diaOD; iz++) {
-                        if (weightOD[oz * diaOD + iz] == 0.f)
-                            continue;
-                        for (int iy = 0; iy < diaOH; iy++) {
-                            if (weightOH[oy * diaOH + iy] == 0.f) {
-                                continue;
-                            }
-                            for (int ix = 0; ix < diaOW; ix++) {
-                                if (weightOW[ox * diaOW + ix] == 0.f) {
-                                    continue;
-                                }
-                                float w = weightOD[oz * diaOD + iz] * weightOH[oy * diaOH + iy] * weightOW[ox * diaOW + ix];
-                                float value = getValue(in_ptr_nc,
-                                    (idxOD[oz * diaOD + iz] * IH * IW + idxOH[oy * diaOH + iy] * IW + idxOW[ox * diaOW + ix]) * srcDataSize, inputPrec);
-
-                                sum += w * value;
-                                wsum += w;
-                            }
-                        }
-                    }
-
-                    if (!wsum) {
-                        setValue(out_ptr_ncdh, ox * dstDataSize, 0.f, outputPrec);
-                    } else {
-                        float dst_value = sum / wsum;
-                        setValue(out_ptr_ncdh, ox * dstDataSize, dst_value, outputPrec);
-                    }
-                }
-            }
-        }
-    });
-}
-
-Interpolate::InterpolateExecutor::InterpolateExecutor(const InterpolateAttrs& interpAttrs,
-                                                                const VectorDims &srcDims,
-                                                                const VectorDims &dstDims,
-                                                                const std::vector<float> &dataScales) :
-        mode(interpAttrs.mode), configured_for_layout(interpAttrs.layout), coordTransMode(interpAttrs.coordTransMode),
-        inputPrec(interpAttrs.inPrc), outputPrec(interpAttrs.outPrc) {
-    srcDimPad5d = to5Dim(getPaddedInputShape(srcDims, interpAttrs.padBegin, interpAttrs.padEnd));
-    dstDim5d = to5Dim(dstDims);
-    srcDataSize = interpAttrs.inPrc.size();
-    dstDataSize = interpAttrs.outPrc.size();
-    dataRank = srcDims.size();
-    spatialDimSize = getSpatialDimsNum(dataRank);
-
-    switch (mode) {
-        case InterpolateMode::nearest: {
-            buildTblNN(srcDimPad5d, dstDim5d, dataScales, interpAttrs.layout, interpAttrs.nearestMode);
-            break;
-        }
-        case InterpolateMode::linear_onnx: {
-            buildTblLinearOnnx(srcDimPad5d, dstDim5d, dataScales, interpAttrs.layout);
-            break;
-        }
-        case InterpolateMode::linear: {
-            static constexpr int LINEAR_KERNEL = 2;
-            buildTblLinear(srcDimPad5d, dstDim5d, dataScales, LINEAR_KERNEL, interpAttrs.antialias);
-            break;
-        }
-        case InterpolateMode::cubic: {
-            buildTblCubic(srcDimPad5d, dstDim5d, dataScales, interpAttrs.cubeCoeff, interpAttrs.layout);
-            break;
-        }
-        default: {
-            IE_THROW() << "Interpolate executor does not support interpolate mode: " << mode;
-            break;
-        }
-    }
-}
-
 Interpolate::InterpolateJitExecutor::InterpolateJitExecutor(const InterpolateAttrs& interpAttrs,
                                                                       const VectorDims &srcDims,
                                                                       const VectorDims &dstDims,
@@ -3201,54 +2407,6 @@ void Interpolate::InterpolateJitExecutor::exec(const uint8_t *in_ptr_, uint8_t *
         default: {
             IE_THROW() << "InterpolateJitExecutor has unsupported interpolate mode: " << mode;
         }
-    }
-}
-
-void Interpolate::InterpolateRefExecutor::exec(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_) {
-    size_t N = srcDimPad5d[0], C = srcDimPad5d[1], ID = srcDimPad5d[2], IH = srcDimPad5d[3], IW = srcDimPad5d[4];
-    size_t OD = dstDim5d[2], OH = dstDim5d[3], OW = dstDim5d[4];
-
-    switch (mode) {
-        case InterpolateMode::nearest: {
-            NNRef(in_ptr_, out_ptr_, N, C, ID, IH, IW, OD, OH, OW);
-            break;
-        }
-        case InterpolateMode::linear_onnx: {
-            linearOnnxRef(in_ptr_, out_ptr_, N, C, ID, IH, IW, OD, OH, OW);
-            break;
-        }
-        case InterpolateMode::cubic: {
-            cubicRef(in_ptr_, out_ptr_, N, C, IH, IW, OH, OW);
-            break;
-        }
-        case InterpolateMode::linear: {
-            float fz = (dataRank == 5) ? dataScales[dataRank - 3] : 1.f;
-            float fy = dataScales[dataRank - 2];
-            float fx = dataScales[dataRank - 1];
-
-            bool isDownsample = (fx < 1.f) || (fy < 1.f) || (fz < 1.f);
-            int kernel_width = 2;
-            linearInterpolation(in_ptr_, out_ptr_, N, C, ID, IH, IW, fx, fy, fz, OD, OH, OW, kernel_width, isDownsample && antialias);
-            break;
-        }
-        default: {
-            IE_THROW() << "Interpolate layer has unsupported interpolate mode: " << mode;
-        }
-    }
-}
-
-size_t Interpolate::getSpatialDimsNum(const Dim rank) {
-    switch (rank) {
-        case 1:
-        case 3:
-            return 1;
-        case 2:
-        case 4:
-            return 2;
-        case 5:
-            return 3;
-        default:
-            IE_THROW() << "Can't define number spatial";
     }
 }
