+++ src/plugins/intel_cpu/src/nodes/interpolate.h
@@ -4,8 +4,7 @@
 
 #pragma once
 
-#include <ie_common.h>
-#include <node.h>
+#include "executors/interpolate.hpp"
 #include <string>
 #include <memory>
 #include <vector>
@@ -17,40 +16,6 @@ using namespace InferenceEngine;
 namespace ov {
 namespace intel_cpu {
 namespace node {
-
-enum InterpolateLayoutType {
-    planar,
-    block,
-    by_channel
-};
-
-enum InterpolateMode {
-    nearest,
-    linear,
-    linear_onnx,
-    cubic
-};
-
-enum InterpolateCoordTransMode {
-    half_pixel,
-    pytorch_half_pixel,
-    asymmetric,
-    tf_half_pixel_for_nn,
-    align_corners
-};
-
-enum class InterpolateNearestMode {
-    round_prefer_floor,
-    round_prefer_ceil,
-    floor,
-    ceil,
-    simple
-};
-
-enum class InterpolateShapeCalcMode {
-    sizes,
-    scales
-};
 
 struct jit_interpolate_config_params {
     InterpolateLayoutType layout;
@@ -121,59 +86,8 @@ public:
     bool needPrepareParams() const override;
     void prepareParams() override;
 
-    struct InterpolateAttrs {
-        InterpolateMode mode = InterpolateMode::nearest;
-        InterpolateCoordTransMode coordTransMode = InterpolateCoordTransMode::half_pixel;
-        InterpolateNearestMode nearestMode = InterpolateNearestMode::round_prefer_floor;
-        bool antialias = false;
-        float cubeCoeff = -0.75;
-        std::vector<int> padBegin;
-        std::vector<int> padEnd;
-        InferenceEngine::Precision inPrc;
-        InferenceEngine::Precision outPrc;
-        InterpolateLayoutType layout;
-    };
-
 private:
     InterpolateAttrs interpAttrs;
-
-    class InterpolateExecutor {
-        public:
-            InterpolateExecutor(const InterpolateAttrs& interpAttrs,
-                                const VectorDims &srcDims,
-                                const VectorDims &dstDims,
-                                const std::vector<float> &dataScales);
-
-            virtual void exec(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_) = 0;
-            virtual ~InterpolateExecutor() = default;
-            VectorDims getSrcDimPad5d() const { return srcDimPad5d; }
-
-        private:
-            void buildTblNN(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d, const std::vector<float>& dataScales,
-                            InterpolateLayoutType layout, InterpolateNearestMode nearestMode);
-            void buildTblLinearOnnx(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d, const std::vector<float>& dataScales,
-                                    InterpolateLayoutType layout);
-            void buildTblLinear(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d, const std::vector<float>& dataScales, int kernel_width,
-                                bool antialias);
-            void buildTblCubic(const SizeVector& srcDimPad5d, const SizeVector& dstDim5d, const std::vector<float>& dataScales, float cubicCoeff,
-                               InterpolateLayoutType layout);
-
-            float coordTransToInput(int outCoord, float scale, int inShape, int outShape) const;
-            int nearestRound(float origin, bool isDownsample, InterpolateNearestMode nearestMode) const;
-            void linearOnnxCF(int outCoord, float scale, int inShape, int outShape, int& index0, int& index1, float& weight0, float& weight1);
-            std::vector<float> getCubicCoeffs(float mantissa, float a);
-
-        protected:
-            InterpolateMode mode;
-            InterpolateCoordTransMode coordTransMode;
-            InterpolateLayoutType configured_for_layout;
-            VectorDims srcDimPad5d, dstDim5d;
-            InferenceEngine::Precision inputPrec, outputPrec;
-            size_t srcDataSize, dstDataSize;
-            int spatialDimSize;
-            size_t dataRank;
-            std::vector<int> indexTable;
-    };
     std::shared_ptr<InterpolateExecutor> execPtr = nullptr;
 
     class InterpolateJitExecutor : public InterpolateExecutor {
@@ -183,7 +97,10 @@ private:
                                    const VectorDims &dstDims,
                                    const std::vector<float> &dataScales,
                                    const dnnl::primitive_attr &attr);
-
+            bool init(const InterpolateAttrs& reduceAttrs,
+                      const std::vector<MemoryDescPtr>& srcDescs,
+                      const std::vector<MemoryDescPtr>& dstDescs,
+                      const dnnl::primitive_attr &attr) override { return true; };
             void exec(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_) override;
 
         private:
@@ -209,37 +126,9 @@ private:
             std::shared_ptr<jit_uni_interpolate_kernel> interpolateKernel = nullptr;
     };
 
-    class InterpolateRefExecutor : public InterpolateExecutor {
-        public:
-            InterpolateRefExecutor(const InterpolateAttrs& interpAttrs,
-                                   const VectorDims &srcDims,
-                                   const VectorDims &dstDims,
-                                   const std::vector<float> &_dataScales) : dataScales(_dataScales), antialias(interpAttrs.antialias),
-                InterpolateExecutor(interpAttrs, srcDims, dstDims, _dataScales) {}
-
-            void exec(const uint8_t *in_ptr_, uint8_t *out_ptr_, const void *post_ops_data_) override;
-
-        private:
-            void NNRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW, int OD, int OH, int OW);
-            void linearOnnxRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW, int OD, int OH, int OW);
-
-            void cubicRef(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int IH, int IW, int OH, int OW);
-            void linearInterpolation(const uint8_t *in_ptr_, uint8_t *out_ptr_, int B, int C, int ID, int IH, int IW,
-                                      float fx, float fy, float fz, int OD, int OH, int OW, int kernel_width, bool antialias);
-
-            static float getValue(const uint8_t *base, size_t offset, InferenceEngine::Precision prec);
-            static void setValue(uint8_t *base, size_t offset, float value, InferenceEngine::Precision prec);
-
-        private:
-            bool antialias;
-            std::vector<float> dataScales;
-    };
-
     void setPostOps(dnnl::primitive_attr &attr, const VectorDims &dims);
 
-    static SizeVector getPaddedInputShape(const VectorDims &srcDims, const std::vector<int> &padBegin, const std::vector<int> &padEnd);
     std::vector<float> getScales(const VectorDims &srcDimPad, const VectorDims &dstDim);
-    static size_t getSpatialDimsNum(const Dim rank);
 
     bool hasPad = false;
     InterpolateShapeCalcMode shapeCalcMode;
