+++ src/plugins/intel_cpu/tests/functional/CMakeLists.txt
@@ -8,7 +8,7 @@ set(TARGET_NAME ov_cpu_func_tests)
 # is a specific version for debugging purpose, just set DEBUG_SRC_PATH
 # to the test case to be debugged and debug using cpuDebugFuncTests
 set(DEBUG_TARGET_NAME cpuDebugFuncTests)
-set(DEBUG_SRC_PATH ${CMAKE_CURRENT_SOURCE_DIR}/single_layer_tests/mvn.cpp)
+set(DEBUG_SRC_PATH ${CMAKE_CURRENT_SOURCE_DIR}/single_layer_tests/interpolate.cpp)
 
 add_library(cpuSpecificRtInfo STATIC $<TARGET_PROPERTY:openvino_intel_cpu_plugin,SOURCE_DIR>/src/utils/rt_info/memory_formats_attribute.hpp
                                      $<TARGET_PROPERTY:openvino_intel_cpu_plugin,SOURCE_DIR>/src/utils/rt_info/memory_formats_attribute.cpp)
